+++ python/mcap/stream_reader.py
